+++ kitty/boss.py
@@ -282,7 +282,7 @@ class DumpCommands:  # {{{
                 if isinstance(x, dict):
                     return json.dumps(x)
                 return x
-            safe_print(what, *map(fmt, a))
+            safe_print(what, *map(fmt, a), flush=True)
 # }}}