+++ docs/changelog.rst
@@ -118,6 +118,8 @@ Detailed list of changes
 - macOS: Fix hiding quick access terminal window not restoring focus to
   previously active application (:disc:`8840`)
 
+- Allow using backspace to move the cursor onto the previous line in cooked mode. This is indicated by the `bw` propert in  kitty's terminfo (:iss:`8841`)
+
 0.42.2 [2025-07-16]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~